move conftest contents to asdf/conftest
